:recycle: Separate file for plural
